+++ src/xulbux/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.9.3"
+__version__ = "1.9.4"
 
 __author__ = "XulbuX"
 __email__ = "xulbux.real@gmail.com"